Get rid of some size_t warnings in OpenMesh
git-svn-id: http://openmesh.org/svnrepo/OpenMesh@839 fdac6126-5c0c-442c-9429-916003d36597

+++ trunk/src/OpenMesh/Core/IO/SR_binary_vector_of_fundamentals.inl
@@ -30,7 +30,7 @@ template <> struct binary< std::vector< T > > {                 \
     size_t bytes=0;                                             \
                                                                 \
     if ( _swap)                                                 \
-      bytes = std::accumulate( _v.begin(), _v.end(), 0,         \
+      bytes = std::accumulate( _v.begin(), _v.end(), size_t(0),         \
 			       FunctorRestore<elem_type>(_is, _swap) );           \
     else                                                        \
     {                                                           \
