+++ trunk/src/OpenMesh/Core/Utils/Property.hh
@@ -135,7 +135,7 @@ public:
   {
     if (element_size() != IO::UnknownSize)
       return this->BaseProperty::size_of(n_elements());
-    return std::accumulate(data_.begin(), data_.end(), 0, plus());
+    return std::accumulate(data_.begin(), data_.end(), size_t(0), plus());
   }
 
   virtual size_t size_of(size_t _n_elem) const
